+++ back-end/package.json
@@ -34,6 +34,7 @@
     "typescript": "^4.9.3"
   },
   "dependencies": {
+    "@types/socket.io": "^3.0.2",
     "config": "^3.3.8",
     "cors": "^2.8.5",
     "dotenv": "^16.0.3",
