+++ back-end/src/routes/Student/StudentRoutes.ts
@@ -9,22 +9,6 @@ import {
 
 const StudentRoute = express.Router()
 
-// const options: cors.CorsOptions = {
-//   allowedHeaders: [
-//     'Origin',
-//     'X-Requested-With',
-//     'Content-Type',
-//     'Accept',
-//     'X-Access-Token'
-//   ],
-//   credentials: true,
-//   methods: 'GET,HEAD,OPTIONS,PUT,PATCH,POST,DELETE',
-//   origin: 'http://localhost:3000',
-//   preflightContinue: false
-// }
-
-// StudentRoute.use(cors(options))
-
 StudentRoute.get('/', getAllStudents)
 StudentRoute.post('/', addStudent)
 StudentRoute.put('/', updateStudent)
