+++ front-end/package.json
@@ -32,6 +32,7 @@
     "react-router-dom": "^6.4.3",
     "react-scripts": "5.0.1",
     "redux-saga": "^1.2.1",
+    "socket.io-client": "^4.5.4",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
